requirements.txt
kaleido==0.2.1
matplotlib==3.7.1
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
plantcv==4.0.1
plotly==5.17.0
tensorflow==2.13.1
